Servant.API.WebSocketConduit

Documentation

data WebSocketConduit i o

Endpoint for defining a route to provide a websocket. In contrast to the WebSocket endpoint, WebSocketConduit provides a higher-level interface. The handler function must be of type Conduit i m o with i and o being instances of FromJSON and ToJSON respectively. await reads from the web socket while yield writes to it.

Example:

import Data.Aeson (Value)
import qualified Data.Conduit.List as CL

type WebSocketApi = "echo" :> WebSocketConduit Value Value

server :: Server WebSocketApi
server = echo
 where
  echo :: Monad m => ConduitT Value Value m ()
  echo = CL.map id

Note that the input format on the web socket is JSON, hence this example only echos valid JSON data.

data WebSocketSource o

Endpoint for defining a route to provide a websocket. In contrast to the WebSocketConduit, this endpoint only produces data. It can be useful when implementing web sockets that simply just send data to clients.

Example:

import Data.Text (Text)
import qualified Data.Conduit.List as CL

type WebSocketApi = "hello" :> WebSocketSource Text

server :: Server WebSocketApi
server = hello
 where
  hello :: Monad m => Conduit Text m ()
  hello = yield $ Just "hello"

runConduitWebSocket :: (MonadBaseControl IO m, MonadUnliftIO m) => Connection -> ConduitT () Void (ResourceT m) () -> m ()

upgradeRequired :: ServerError